A lot happened. see desc
Fixed classes on the homepage (for the like 8th time), and added the day and another schedule to the schedule, and also made sweeping changes to how the schedule editor works, adding support for templates and reworking the whole way that schedule  interacts with the database. There's more too, probably. Ooh and I remembered to lint and format it!