Show time on classes, sort classes by time (#43)
* Update index.tsx

* Draggable UI with Icons

Cut down on useless things, general clean up, made things better, this is a bad commit message but v3 of this should be way better so who cares

* General Clean up

Uses types in dataTransfer instead of useState to do some stuff

* Finished making draggable components

* linted, and remembered to remove old stuff

* Preview state is now optional, left/right UI swapping

* Added styling support for zones, removed old comments

* Removed previews, which prevents the seizure problem

* Shows time in the pill, sorts by time

* Update schedule.tsx

* fixed classes mixing up time with special events

* fixes idk at this point

* fixes

* honestly no idea what just happened

* .

* revert loading page

* remove 3x2 dot moving icons

* merge old loading page back in

* help me God

* Made it actually work

deleted several functions that weren't in use or were doing something incorrectly.

* linted and formatted, added

* Fixed sorting function

* linted + formatted

* A lot happened. see desc

Fixed classes on the homepage (for the like 8th time), and added the day and another schedule to the schedule, and also made sweeping changes to how the schedule editor works, adding support for templates and reworking the whole way that schedule  interacts with the database. There's more too, probably. Ooh and I remembered to lint and format it!

* removed some console.logs that I forgot about

* pretty sure I resolved merge conflicts?

* MLH

* fixed flickering on firefox

Class images still flicker on Safari

* fixed image flickering on safari

* reduced fetch requests by 100,000 %

---------

Co-authored-by: Lukas <69743171+quick007@users.noreply.github.com>